[HOPSWORKS-214] have only one livy user and one notebookServer per project.
