+++ hopsworks-api/src/main/java/io/hops/hopsworks/api/zeppelin/rest/ConfigurationsRestApi.java
@@ -89,8 +89,7 @@ public class ConfigurationsRestApi {
       throw new AppException(Response.Status.FORBIDDEN.getStatusCode(),
               "You curently have no role in this project!");
     }
-    ZeppelinConfig zeppelinConf = zeppelinConfFactory.getZeppelinConfig(project.
-            getName(), user.getEmail());
+    ZeppelinConfig zeppelinConf = zeppelinConfFactory.getProjectConf(project.getName());
     if (zeppelinConf == null) {
       throw new AppException(Response.Status.BAD_REQUEST.getStatusCode(),
               "Could not connect to web socket.");
@@ -126,8 +125,7 @@ public class ConfigurationsRestApi {
       throw new AppException(Response.Status.FORBIDDEN.getStatusCode(),
               "You curently have no role in this project!");
     }
-    ZeppelinConfig zeppelinConf = zeppelinConfFactory.getZeppelinConfig(project.
-            getName(), user.getEmail());
+    ZeppelinConfig zeppelinConf = zeppelinConfFactory.getProjectConf(project.getName());
     if (zeppelinConf == null) {
       throw new AppException(Response.Status.BAD_REQUEST.getStatusCode(),
               "Could not connect to web socket.");
