+++ hopsworks-api/src/main/java/io/hops/hopsworks/api/zeppelin/rest/NotebookService.java
@@ -69,8 +69,7 @@ public class NotebookService {
               "You curently have no role in this project!");
     }
 
-    ZeppelinConfig zeppelinConf = zeppelinConfFactory.getZeppelinConfig(project.
-            getName(), user.getEmail());
+    ZeppelinConfig zeppelinConf = zeppelinConfFactory.getProjectConf(project.getName());
     if (zeppelinConf == null) {
       throw new AppException(Response.Status.BAD_REQUEST.getStatusCode(),
               "Could not connect to web socket.");
@@ -80,7 +79,7 @@ public class NotebookService {
       zeppelinConf.getNotebookRepo().list(AuthenticationInfo.ANONYMOUS);
     }
 
-    notebookRestApi.setParms(project, userRole, hdfsController.getHdfsUserName(
+    notebookRestApi.setParms(project, hdfsController.getHdfsUserName(
             project, user), zeppelinConf);
     return notebookRestApi;
   }
